backend/requirements.txt
fastapi>=0.136.1,<0.137
uvicorn[standard]>=0.35.0,<1
sqlalchemy>=2.0.43,<2.1
pydantic-settings>=2.14.0,<3
python-multipart>=0.0.20,<0.1
opencv-python-headless>=4.13.0.92,<5
imagehash>=4.3.2,<5
numpy>=2.4.4,<3
scipy>=1.17.1,<2
Pillow>=11.3.0,<12
scikit-image>=0.26.0,<0.27
psycopg[binary]>=3.2.9,<4
httpx>=0.28.1,<0.29
pytest>=8.4.1,<9

